backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
google-genai>=1.0.0
python-dotenv>=1.0.0
